Avoid flicker on every query change.
This is the best fix I've got so far to prevent that flicker every keypress.
No more clearing of suggestions on every change of query.  Just return
an empty suggestions list to the callback when getting the suggestions
with a query shorter than the minimum length so that the suggestions
list will be empty when rendered.